Consolidate tests that cover reading from RC file
Authored-by: Owen Nelson <owen.nelson@thoughtworks.com>
